+++ src/checkouts/types.ts
@@ -1,6 +1,6 @@
 import z from "zod";
-import { YunoAmount, YunoMetadata } from "../shared/types/common";
-import { amountSchema, metadataSchema } from "../shared/types";
+import { YunoAmount, YunoMetadata, YunoCard, YunoCardData, YunoBrowserInfo } from "../shared/types/common";
+import { amountSchema, metadataSchema, cardDataSchema, browserInfoSchema } from "../shared/types";
 
 export interface YunoInstallmentsPlan {
   installment: number;
@@ -38,30 +38,6 @@ export interface YunoCheckoutPaymentMethodsResponse {
   payment_methods: YunoCheckoutPaymentMethod[];
 }
 
-export interface YunoBrowserInfo {
-  browser_time_difference: string;
-  color_depth: string;
-  java_enabled: boolean;
-  screen_width: string;
-  screen_height: string;
-  user_agent: string;
-  language: string;
-  javascript_enabled: boolean;
-  accept_browser: string;
-  accept_content: string;
-  accept_header: string;
-}
-
-export interface YunoCard {
-  expiration_month: number;
-  expiration_year: number;
-  number: string;
-  security_code: string;
-  holder_name: string;
-  type?: string | null;
-  brand?: string;
-}
-
 export interface YunoOttCustomer {
   browser_info: YunoBrowserInfo;
 }
@@ -83,19 +59,6 @@ export interface YunoOttRequest {
   installment?: any | null;
   third_party_data?: any | null;
   device_fingerprints?: any | null;
-}
-
-export interface YunoCardData {
-  holder_name: string;
-  iin: string;
-  lfd: string;
-  number_length: number;
-  security_code_length: number;
-  brand: string;
-  type: string;
-  category: string;
-  issuer_name: string;
-  issuer_code?: string | null;
 }
 
 export interface YunoOttResponseCustomer {
@@ -151,35 +114,19 @@ export const checkoutSessionCreateSchema = z.object({
     .describe("The installment plan configuration"),
 });
 
-// Schema para crear OTT
+// Schema para crear OTT - usando schemas compartidos
 export const ottCreateSchema = z.object({
   sessionId: z.string().describe("The unique identifier of the checkout session"),
   payment_method: z.object({
     type: z.string().describe("Payment method type (e.g., 'CARD')"),
     vault_on_success: z.boolean().describe("Whether to vault the payment method on success"),
-    card: z.object({
-      expiration_month: z.number().int().min(1).max(12).describe("Card expiration month (1-12)"),
+    card: cardDataSchema.extend({
       expiration_year: z.number().int().min(20).max(99).describe("Card expiration year (YY format)"),
-      number: z.string().describe("Card number"),
       security_code: z.string().describe("Card security code (CVV)"),
       holder_name: z.string().describe("Cardholder name"),
-      type: z.string().optional().nullable().describe("Card type"),
-      brand: z.string().optional().describe("Card brand (e.g., 'VISA', 'MASTERCARD')"),
     }),
     customer: z.object({
-      browser_info: z.object({
-        browser_time_difference: z.string().describe("Browser time difference"),
-        color_depth: z.string().describe("Screen color depth"),
-        java_enabled: z.boolean().describe("Whether Java is enabled"),
-        screen_width: z.string().describe("Screen width"),
-        screen_height: z.string().describe("Screen height"),
-        user_agent: z.string().describe("Browser user agent"),
-        language: z.string().describe("Browser language"),
-        javascript_enabled: z.boolean().describe("Whether JavaScript is enabled"),
-        accept_browser: z.string().describe("Browser accept header"),
-        accept_content: z.string().describe("Content accept header"),
-        accept_header: z.string().describe("Accept header"),
-      }),
+      browser_info: browserInfoSchema,
     }),
   }),
   three_d_secure: z.object({
